+++ reactionmenu/abc.py
@@ -945,6 +945,38 @@ class BaseMenu(metaclass=abc.ABCMeta):
                     else:
                         raise MenuException(f'When using parameter "send_to" in {class_name}.start(), the channel {send_to} was not found')
 
+    def set_page_director_style(self, style_id: int) -> None:
+        """Set how the page numbers dictating what page you are one (in the footer of an embed/regular message) are displayed
+
+        Parameters
+        ----------
+        style_id: :class:`int`
+            Varying formats of how the page director can be presented. The following ID's are available:
+
+            - `1` = Page 1/10
+            - `2` = Page 1 out of 10
+            - `3` = 1 • 10
+            - `4` = 1 » 10
+            - `5` = 1 | 10
+            - `6` = 1 : 10
+            - `7` = 1 - 10
+            - `8` = 1 / 10
+        
+        Raises
+        ------
+        - `MenuException`: The :param:`style_id` value was not valid 
+        """
+        if style_id == 1:   self.style = _DEFAULT_STYLE
+        elif style_id == 2: self.style = 'Page $ out of &'
+        elif style_id == 3: self.style = '$ • &'
+        elif style_id == 4: self.style = '$ » &'
+        elif style_id == 5: self.style = '$ | &'
+        elif style_id == 6: self.style = '$ : &'
+        elif style_id == 7: self.style = '$ - &'
+        elif style_id == 8: self.style = '$ / &'
+        else:
+            raise MenuException(f'Parameter "style_id" expected a number 1-8, got {style_id!r}')
+    
     @ensure_not_primed
     async def add_from_messages(self, messages: Sequence[discord.Message]) -> None:
         """|coro|